EID-625 Use library version of eIDAS LoA mapping
- Also remove a level of indirection of classes.
